Move YAML-relevant files to a uniform package
