+++ src/main/resources/META-INF/plugin.xml
@@ -166,11 +166,11 @@
 
         <completion.confidence language="Arend" implementationClass="org.arend.codeInsight.completion.ArendCompletionConfidence" id="arendCompletion" />
 
-        <completion.contributor language="yaml" implementationClass="org.arend.codeInsight.completion.YAMLCompletionContributor"/>
+        <completion.contributor language="yaml" implementationClass="org.arend.yaml.codeInsight.YAMLCompletionContributor"/>
 
         <!-- Annotation -->
 
-        <annotator language="yaml" implementationClass="org.arend.annotation.YAMLHighlightingAnnotator"/>
+        <annotator language="yaml" implementationClass="org.arend.yaml.codeInsight.YAMLHighlightingAnnotator"/>
 
         <!-- Intention Actions -->
         <intentionAction>
